stable iso refactor of docs routes
